Move source files.
